% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refs.R
\name{lk_ontoProc}
\alias{lk_ontoProc}
\title{ontoProc: convenient representations of ontologies for genomic data science}
\usage{
lk_ontoProc()
}
\description{
ontoProc: convenient representations of ontologies for genomic data science
}
\examples{
if (requireNamespace("ontoProc")) {
 cl = ontoProc::getOnto('cellOnto')
 cl3k = c("CL:0000492", "CL:0001054", "CL:0000236", "CL:0000625",
    "CL:0000576", "CL:0000623", "CL:0000451", "CL:0000556")
 ontoProc::onto_plot2(cl, cl3k)
}
}
